Give algorithm the correct name
